requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.0
better-profanity==0.7.0
slowapi==0.1.9
jinja2==3.1.3
python-multipart==0.0.6
pydantic-settings==2.1.0
alembic==1.13.1
pytest==7.4.3
httpx==0.26.0
